fix(test): use correct param and equivalence check
This test had never actually been making the correct assertion here as
it was doing an always true `if assignableTopic == assignableTopic`
check. After fixing that it became clear that the
Test_stickyBalanceStrategy_Plan_ReassignmentWithRandomSubscriptionsAndChanges
test was passing the wrong members parameter for its second phase. After
fixing both of those the test passed and checked the correct state.