feat: allow balance strategies to provide initial state (#1633)
* Allow balance strategies to provide initial state

This update will allow stateful balance strategies to provide initial state information to send as part of the `joinGroupRequest`.

Previous to this PR if initial data was provided via `c.config.Consumer.Group.Member.UserData` it would always be sent when re-joining. This would trump the previous assignment data returned as part of the `syncGroupRequest`. After this PR the initial data will be sent when there is no currently assigned user-data that was returned as part of the assignment.

Co-authored-by: Vlad Gorodetsky <v@gor.io>