feat: send ApiVersionsRequest on broker open
Currently we don't use the actual response for anything, but lay the
groundwork by sending an ApiVersionsRequest on broker open when the
kafka version config has been set to 2.4.0.0 or newer. This is useful
because the clientSoftwareName and version will show up in the
server-side metrics.

```
kafka.server:clientSoftwareName=sarama,clientSoftwareVersion=1.30.0,listener=PLAINTEXT,networkProcessor=1,type=socket-server-metrics
```