fix(consumer): default ResetInvalidOffsets to true
A number of people have reported issues with the change in default
behaviour that was introduced in #2252 and the comment for this
configuration value had suggested it would retain the behaviour of
previous Sarama versions. This PR puts the default behaviour back to
resetting to `c.Consumer.Offsets.Initial` when the server returns an out
of range error on a Fetch.

Contributes-to: #2342