fix: admin retry logic
The previous fix wasn't correct. While looking at the transaction_manager.go
logic, it became clear that the intention is that retry.max should not
count the first attempt. This seems very reasonable so this fixes the test
and logic to reflect that. The original code would
fail this updated test because it didn't make the final required attempt.

Signed-off-by: Mark Hindess <mark.hindess@gmail.com>