chore(ci): replace toxiproxy client dep
The v2.5.0 client implementation happens to have a resource leak, and
pulling in HEAD notably carries a large number of transitive
dependencies that we don't really want to track in our go.mod just for
FVT purposes. Therefore, knock up a minimal implementation of an http
client for toxiproxy to satisfy our usage.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>