chore(ci): add goleak to TestMain
Whilst this isn't as helpful as a per test `defer goleak.VerifyNone` in
terms of showing _which_ test was responsible for the leak, it's still
worth having as a baseline check in PRs and CI to prevent new leaks
being introduced now that we're all cleaned up. If a PR fails we can
subsequently run tests individually in order to determine which one is
the leaker.

See https://github.com/uber-go/goleak#quick-start and
https://github.com/uber-go/goleak#determine-source-of-package-leaks

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>