Fix data race on Broker.done channel
The underlying case was not waiting for the goroutine running the
`responseReceiver()` method to fully complete if SASL authentication
failed. This created a window where a further call to `Broker.Open()`
could overwrite the `Broker.done` channel value while the goroutine
still running `responseReceiver()` was trying to close the same channel.

Fixes: #2382
Signed-off-by: Adrian Preston <PRESTONA@uk.ibm.com>