fix: return ErrMessageSizeTooLarge when message is too large
For most of this library's existence it has returned ErrMessageSizeTooLarge
when the message exceeded the configured size.

For a short period in 2019 this error was renamed (#1218) but shortly
revered back (#1262). Later in 2023 this error was changed to a
ConfigurationError (#2628) to fix #2137, however this has caused issues
with clients who rely on the previous error code being distinct from
other ConfigurationError conditions (#2655).

This commit reverts to previous behaviour, and adds a test to pickup if
this changes again in the future.

Signed-off-by: Adam Eijdenberg <adam.eijdenberg@defence.gov.au>