fix(decoder): handle null arrays correctly
Array lengths in the Kafka protocol are encoded as a 32-bit signed
integer in big-endian format. For a null value this corresponds to the
byte sequence FF FF FF FF which is intended to represent null as a
length of -1. To correctly read that we need to cast the value to int32
before widening to int.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>