fix(consumer): stuck on the batch with zero records length (#3221)
This pull request addresses the known issue described in
[KAFKA-5443](https://issues.apache.org/jira/browse/KAFKA-5443), which
was initially analyzed in
[#2053](https://github.com/IBM/sarama/pull/2053). A prior attempt to fix
it was made in [#2057](https://github.com/IBM/sarama/pull/2057), but the
fix was applied incorrectly and did not fully resolve the root cause.

I’ve provided a more detailed explanation of the underlying issue and
the reasoning behind this fix in [this
comment](https://github.com/IBM/sarama/pull/3150#issuecomment-3064757832).

---------

Signed-off-by: Sterligov Denis <d.sterligov@tinkoff.ru>
Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>
Co-authored-by: Sterligov Denis <d.sterligov@tinkoff.ru>
Co-authored-by: Dominic Evans <dominic.evans@uk.ibm.com>