feat: always send ApiVersionsRequest and fallback to v0 (#3234)
We previously pinned to only sending if Version was set to v2.4.0.0, but
we should always send the request regardless, but fallback to sending a
v0 request (without client info) if the v3 request does not succeed

In functional tests we always use MaxVersion in config an
ApiVersionsRequest enabled to negotiate versions.

In unittests we disable ApiVersionsRequest by default and pin Version. A
number of these tests use step-by-step ordered protocol expectations and
it isn't convenient to update them all to also include an
ApiVersionsRequest roundtrip unnecessarily, so we disable it in plain
unittests by default, keeping it as opt-in.

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>
Co-authored-by: Giorgio Pellero <giorgio.pellero@typeform.com>