feat: add SASLv1 support for Kerberos (#3279)
For some reason we were still pinning the old v0 SASL when using GSSAPI
and this doesn't work if an ApiVersionsRequest is sent before the auth
flow.
Add support for sending the krb5 bytes in the v1 SASL SaslAuthenticate
protocol wrapping.

Note: this is just a draft for now as although it is tested and working
against a GSSAPI enabled cluster, the underlying code is very much a
copy-and-paste of the existing v0 auth flow, just wrappering the bytes
in the SaslAuthenticate rather than sending them to the broker directly.
This needs more work to tidyup the implementation

---------

Signed-off-by: Dominic Evans <dominic.evans@uk.ibm.com>